long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5715 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/export/usb.h
@@ -23,7 +23,7 @@
 
 void usb_init(void);
 void usb_start_monitoring(void);
-void usb_acknowledge(int id);
+void usb_acknowledge(long id);
 void usb_wait_for_disconnect(struct event_queue *q);
 int usb_wait_for_disconnect_w_tmo(struct event_queue *q, int ticks);
 bool usb_inserted(void); /* return the official value, what's been reported to the threads */
